catalog_engine/Rest-api/requirements.txt
psycopg2-binary==2.9.3
elasticsearch==8.12.0
replicate
sentence_transformers
flask
confluent_kafka==1.9.0